// src/generated/zoom-meetings.d.ts
export interface paths {
    "/meetings/{meetingId}/invite_links": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        /**
         * Create a meeting's invite links
         * @description Create a batch of invitation links for a meeting.
         *
         *     **Prerequisites**:
         *     * The `ttl` value, in seconds, defines the invite link's expiration time. It must be between `0` or no expiration and `7776000` or 90 days, and has a default value of `7200` or 2 hours.
         *
         *     **[Scopes](https://developers.zoom.us/docs/integrations/oauth-scopes-overview/):** `meeting:write:admin`,`meeting:write`
         *
         *     **[Granular Scopes](https://developers.zoom.us/docs/integrations/oauth-scopes-overview/):** `meeting:write:invite_links`,`meeting:write:invite_links:admin`
         *
         *     **[Rate Limit Label](https://marketplace.zoom.us/docs/api-reference/rate-limits#rate-limits):** `LIGHT`
         */
        post: operations["meetingInviteLinksCreate"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: never;
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}
export type $defs = Record<string, never>;
export interface operations {
    meetingInviteLinksCreate: {
        parameters: {
            query?: never;
            header?: never;
            path: {
                /**
                 * @description The meeting's ID.
                 *
                 *      When storing this value in your database, you must store it as a long format integer and **not** an integer. Meeting IDs can exceed 10 digits.
                 */
                meetingId: number;
            };
            cookie?: never;
        };
        requestBody?: {
            content: {
                "application/json": {
                    /** @description The attendees list. */
                    attendees?: {
                        /**
                         * @description Whether to disable participant audio when joining the meeting. If not provided or set to `false`, the participant audio will follow the meeting's default settings.
                         * @default false
                         * @example false
                         */
                        disable_audio?: boolean;
                        /**
                         * @description Whether to disable participant video when joining the meeting. If not provided or set to `false`, the participant video will follow the meeting's default settings.
                         * @default false
                         * @example false
                         */
                        disable_video?: boolean;
                        /**
                         * @description User display name.
                         * @example Jill Chill
                         */
                        name: string;
                    }[];
                    /**
                     * Format: int64
                     * @description The invite link's expiration time, in seconds.
                     *
                     *     This value defaults to `7200`.
                     * @default 7200
                     * @example 1000
                     */
                    ttl?: number;
                };
            };
        };
        responses: {
            /**
             * @description **HTTP Status Code:** `201`
             *
             *     Meeting invitation links created.
             */
            201: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": {
                        /** @description The attendee list. */
                        attendees?: {
                            /**
                             * @description The URL to join the meeting.
                             * @example https://example.com/j/11111
                             */
                            join_url?: string;
                            /**
                             * @description The user's display name.
                             * @example Jill Chill
                             */
                            name?: string;
                        }[];
                    };
                };
            };
            /**
             * @description **HTTP Status Code:** `400` <br>
             *      Bad Request
             *
             *      **Error Code:** `300` <br>
             *      Invalid meeting ID. <br>
             *     **Error Code:** `3000` <br>
             *      Cannot access webinar information. <br>
             *     **Error Code:** `3161` <br>
             *      Meeting hosting and scheduling capabilities are not allowed for your user account. <br>
             */
            400: {
                headers: {
                    [name: string]: unknown;
                };
                content?: never;
            };
            /**
             * @description **HTTP Status Code:** `404` <br>
             *      Not Found
             *
             *      **Error Code:** `3001` <br>
             *      Meeting does not exist: {meetingId} <br>
             */
            404: {
                headers: {
                    [name: string]: unknown;
                };
                content?: never;
            };
            /**
             * @description **HTTP Status Code:** `429` <br>
             *      Too Many Requests. For more information, see [rate limits](/docs/api/rest/rate-limits/).
             */
            429: {
                headers: {
                    [name: string]: unknown;
                };
                content?: never;
            };
        };
    };
}